Begin refactor to share code between proxy and reverse_proxy.
