+++ lib/proxy/mutators/oauth_param_collector.js
@@ -1,6 +1,6 @@
  var encoding = require('../../encoding.js');
 
- var oauth_constants = require('../oauth_constants.js');
+ var oauth_constants = require('../oauth/constants.js');
 
 /**
  * This mutator is responsible for scouring an inbound request and gathering OAuth parameters.  These
