+++ lib/proxy/validators/oauth_param_sanity_validator.js
@@ -1,7 +1,7 @@
 var bad_request = require('../messages/bad_request.js');
 var unauthorized = require('../messages/unauthorized.js');
 
-var oauth_constants = require('../oauth_constants.js');
+var oauth_constants = require('../oauth/constants.js');
 
 module.exports = function(proxy) {
