+++ lib/proxy/validators/oauth_timestamp_validator.js
@@ -1,7 +1,7 @@
 var unauthorized = require('../messages/unauthorized.js');
 
-var OAUTH_TIMESTAMP = require('../oauth_constants.js').OAUTH_TIMESTAMP;
-var OAUTH_TIMESTAMP_MAX_AGE = require('../oauth_constants.js').OAUTH_TIMESTAMP_MAX_AGE;
+var OAUTH_TIMESTAMP = require('../oauth/constants.js').OAUTH_TIMESTAMP;
+var OAUTH_TIMESTAMP_MAX_AGE = require('../oauth/constants.js').OAUTH_TIMESTAMP_MAX_AGE;
 
 module.exports = function(proxy) {
   return function(req, res, next) {
