+++ lib/proxy/validators/quota_validator.js
@@ -1,4 +1,4 @@
-var oauth_constants = require('../oauth_constants.js');
+var oauth_constants = require('../oauth/constants.js');
 
 var unauthorized = require('../messages/unauthorized.js');
