+++ lib/proxy/validators/url_length_validator.js
@@ -1,4 +1,4 @@
-var MAXIMUM_URL_LENGTH = require('../oauth_constants.js').MAXIMUM_URL_LENGTH;
+var MAXIMUM_URL_LENGTH = require('../oauth/constants.js').MAXIMUM_URL_LENGTH;
 
 module.exports = function() {
   return function(req, res, next) {
