+++ test/server/test_server.js
@@ -13,7 +13,7 @@ app.POST = app.post;
 app.PUT = app.put;
 app.DELETE = app.delete;
 
-const CONSUMER_KEY_HEADER = require('../../lib/proxy/oauth_constants.js').CONSUMER_KEY_HEADER;
+const CONSUMER_KEY_HEADER = require('../../lib/proxy/oauth/constants.js').CONSUMER_KEY_HEADER;
 
 // The job server represents they types of servers we might expect to see behind oauth_reverse_proxy.
 function JobServer() {
